requirements.txt
ddgs
requests
Pillow
